Potions and Damage Text
Added Button class for the player to click on Potion button to heal.

The bandit's first action is to check if they must heal. This triggers if their health is below half of their max hp and they have potions available.

Added Damage Text class to display attack damage in red above the target or Player after an attack occurs. Also, added heal text in green to display when the player or a bandit heals.